change cli name to 'bombard'
