+++ hw-3/server/bombard/.editorconfig
@@ -4,8 +4,8 @@ root = true
 indent_style = space
 indent_size = 2
 charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
+# trim_trailing_whitespace = true
+# insert_final_newline = true
 
 [*.md]
 trim_trailing_whitespace = false
