+++ hw-3/server/bombard/.gitignore
@@ -5,5 +5,6 @@
 /lib
 /package-lock.json
 /tmp
+/yarn.lock
 node_modules
 oclif.manifest.json
