--- hw-3/src/commands/hello/world.ts
@@ -1,19 +1,0 @@
-import {Command} from '@oclif/core'
-
-export default class World extends Command {
-  static description = 'Say hello world'
-
-  static examples = [
-    `$ oex hello world
-hello world! (./src/commands/hello/world.ts)
-`,
-  ]
-
-  static flags = {}
-
-  static args = []
-
-  async run(): Promise<void> {
-    this.log('hello world! (./src/commands/hello/world.ts)')
-  }
-}
